// vhome_mis/src/app/core/api-client-v2/model/managementServiceUserResponseServiceUserData.ts
/**
 * APIFlask
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ManagementServiceUserResponseServiceUserNickName } from './managementServiceUserResponseServiceUserNickName';
import { ManagementServiceUserResponseLivingSpaceData } from './managementServiceUserResponseLivingSpaceData';


export interface ManagementServiceUserResponseServiceUserData { 
    ID: number;
    active: boolean;
    deleted: boolean;
    living_space: ManagementServiceUserResponseLivingSpaceData | null;
    name: string;
    service_user_nicknames: Array<ManagementServiceUserResponseServiceUserNickName>;
    start_date: string;
}