// vhome_mis/src/app/core/api-client-v2/model/managementServiceUserResponseServiceUserNickName.ts
/**
 * APIFlask
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ManagementServiceUserResponseServiceUserNickName { 
    ID: number;
    nickname: string;
    service_user_id: number;
}